refactor: user 관련 코드 패키지 구조 변경
