+++ src/main/java/com/recipe/app/src/fridge/models/Fridge.java
@@ -2,7 +2,7 @@ package com.recipe.app.src.fridge.models;
 
 import com.recipe.app.common.entity.BaseEntity;
 import com.recipe.app.src.ingredientCategory.models.IngredientCategory;
-import com.recipe.app.src.user.models.User;
+import com.recipe.app.src.user.domain.User;
 import lombok.AccessLevel;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
