+++ src/main/java/com/recipe/app/src/fridgeBasket/models/FridgeBasket.java
@@ -3,7 +3,7 @@ package com.recipe.app.src.fridgeBasket.models;
 import com.recipe.app.common.entity.BaseEntity;
 import com.recipe.app.src.ingredient.models.Ingredient;
 import com.recipe.app.src.ingredientCategory.models.IngredientCategory;
-import com.recipe.app.src.user.models.User;
+import com.recipe.app.src.user.domain.User;
 import lombok.AccessLevel;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
