+++ src/main/java/com/recipe/app/src/recipeInfo/RecipeInfoProvider.java
@@ -12,8 +12,8 @@ import com.recipe.app.src.recipeKeyword.models.RecipeKeyword;
 import com.recipe.app.src.recipeProcess.models.RecipeProcess;
 import com.recipe.app.src.scrapBlog.ScrapBlogRepository;
 import com.recipe.app.src.scrapPublic.ScrapPublicRepository;
-import com.recipe.app.src.user.UserProvider;
-import com.recipe.app.src.user.models.User;
+import com.recipe.app.src.user.application.UserProvider;
+import com.recipe.app.src.user.domain.User;
 import com.recipe.app.common.utils.JwtService;
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
