+++ src/main/java/com/recipe/app/src/scrapBlog/ScrapBlogService.java
@@ -2,8 +2,8 @@ package com.recipe.app.src.scrapBlog;
 
 import com.recipe.app.common.exception.BaseException;
 import com.recipe.app.src.scrapBlog.models.*;
-import com.recipe.app.src.user.UserProvider;
-import com.recipe.app.src.user.models.User;
+import com.recipe.app.src.user.application.UserProvider;
+import com.recipe.app.src.user.domain.User;
 import com.recipe.app.common.utils.JwtService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
