+++ src/main/java/com/recipe/app/src/scrapBlog/models/ScrapBlog.java
@@ -1,6 +1,6 @@
 package com.recipe.app.src.scrapBlog.models;
 
-import com.recipe.app.src.user.models.User;
+import com.recipe.app.src.user.domain.User;
 import lombok.*;
 import javax.persistence.*;
 import com.recipe.app.common.entity.BaseEntity;
