+++ src/main/java/com/recipe/app/src/scrapPublic/ScrapPublicService.java
@@ -4,16 +4,13 @@ import com.recipe.app.src.recipeInfo.RecipeInfoProvider;
 import com.recipe.app.src.recipeInfo.models.RecipeInfo;
 import com.recipe.app.src.scrapPublic.models.PostScrapPublicRes;
 import com.recipe.app.src.scrapPublic.models.ScrapPublic;
-import com.recipe.app.src.user.UserProvider;
+import com.recipe.app.src.user.application.UserProvider;
 import com.recipe.app.common.utils.JwtService;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import com.recipe.app.common.exception.BaseException;
-import com.recipe.app.src.user.models.User;
+import com.recipe.app.src.user.domain.User;
 import org.springframework.stereotype.Service;
-
-
-import static com.recipe.app.common.response.BaseResponseStatus.*;
 
 
 @Service
