+++ src/main/java/com/recipe/app/src/scrapYoutube/ScrapYoutubeProvider.java
@@ -5,8 +5,8 @@ import com.recipe.app.common.exception.BaseException;
 import com.recipe.app.src.scrapYoutube.models.GetScrapYoutubesRes;
 import com.recipe.app.src.scrapYoutube.models.ScrapYoutube;
 import com.recipe.app.src.scrapYoutube.models.ScrapYoutubeList;
-import com.recipe.app.src.user.UserProvider;
-import com.recipe.app.src.user.models.User;
+import com.recipe.app.src.user.application.UserProvider;
+import com.recipe.app.src.user.domain.User;
 import com.recipe.app.src.viewYoutube.ViewYoutubeRepository;
 import com.recipe.app.common.utils.JwtService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,8 +15,6 @@ import org.springframework.stereotype.Service;
 
 import java.util.*;
 import java.util.stream.Collectors;
-
-import static com.recipe.app.common.response.BaseResponseStatus.*;
 
 
 @Service
