+++ src/main/java/com/recipe/app/src/scrapYoutube/ScrapYoutubeService.java
@@ -2,15 +2,11 @@ package com.recipe.app.src.scrapYoutube;
 
 import com.recipe.app.common.exception.BaseException;
 import com.recipe.app.src.scrapYoutube.models.*;
-import com.recipe.app.src.user.UserProvider;
-import com.recipe.app.src.user.models.User;
+import com.recipe.app.src.user.application.UserProvider;
+import com.recipe.app.src.user.domain.User;
 import com.recipe.app.common.utils.JwtService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
-
-
-import static com.recipe.app.common.response.BaseResponseStatus.*;
 
 
 @Service
