+++ src/main/java/com/recipe/app/src/user/application/dto/PatchUserRes.java
@@ -1,4 +1,4 @@
-package com.recipe.app.src.user.models;
+package com.recipe.app.src.user.application.dto;
 
 import lombok.*;
