+++ src/main/java/com/recipe/app/src/viewBlog/ViewBlogService.java
@@ -2,13 +2,10 @@ package com.recipe.app.src.viewBlog;
 
 import com.recipe.app.common.exception.BaseException;
 import com.recipe.app.src.viewBlog.models.*;
-import com.recipe.app.src.user.UserProvider;
-import com.recipe.app.src.user.models.User;
+import com.recipe.app.src.user.application.UserProvider;
+import com.recipe.app.src.user.domain.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
-
-import static com.recipe.app.common.response.BaseResponseStatus.*;
 
 
 @Service
