+++ src/main/java/com/recipe/app/src/viewPublic/models/ViewPublic.java
@@ -2,7 +2,7 @@ package com.recipe.app.src.viewPublic.models;
 
 import com.recipe.app.common.entity.BaseEntity;
 import com.recipe.app.src.recipeInfo.models.RecipeInfo;
-import com.recipe.app.src.user.models.User;
+import com.recipe.app.src.user.domain.User;
 import lombok.AccessLevel;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
