+++ src/main/java/com/recipe/app/src/viewYoutube/ViewYoutubeService.java
@@ -1,8 +1,8 @@
 package com.recipe.app.src.viewYoutube;
 
 import com.recipe.app.common.exception.BaseException;
-import com.recipe.app.src.user.UserProvider;
-import com.recipe.app.src.user.models.User;
+import com.recipe.app.src.user.application.UserProvider;
+import com.recipe.app.src.user.domain.User;
 import com.recipe.app.src.viewYoutube.models.PostViewsYoutubeReq;
 import com.recipe.app.src.viewYoutube.models.PostViewsYoutubeRes;
 import com.recipe.app.src.viewYoutube.models.ViewYoutube;
